Play slider moves with video.

+++ src/gui/Timeline.h
@@ -4,11 +4,11 @@
 class Timeline {
 public:
 	// Constructor.
-	Timeline(int x, int y, float w, float h, const ofColor& color, int numNotches);
+	Timeline(float x, float y, float w, float h, const ofColor& color, int numNotches);
 
 	// Getter methods. 
-	int getX();
-	int getY();
+	float getX();
+	float getY();
 	float width();
 	float height();
 	const ofColor& getColor();
@@ -17,18 +17,18 @@ public:
 	const ofPoint& getPlaySliderPosition();
 
 	// Setter methods.
-	void setX(int);
-	void setY(int);
+	void setX(float);
+	void setY(float);
 	void width(float);
 	void height(float);
 	void setColor(const ofColor&);
 	void setNumNotches(int);
-	void setPlaySliderPosition(int, int);
+	void setPlaySliderPosition(float, float);
 
 	// Draws the timeline.
 	void draw();
 private:
-	int x, y;
+	float x, y;
 	float w, h;
 	ofColor color;
 	int numNotches;
